better separation of responsibilities
